// fresh/weary-crane/types/TerrainType.ts
export type TerrainType =
  | "forest"
  | "sea"
  | "sea-stack"
  | "beach"
  | "river"
  | "lake"
  | "plain"
  | "mountains"
  | "hq"
  | "base-land"
  | "base-air"
  | "base-sea"
  | "power-plant"
  | "road-east-west"
  | "road-north-south"
  | "road-north"
  | "road-east"
  | "road-south"
  | "road-west"
  | "road-north-west"
  | "road-north-east"
  | "road-south-west"
  | "road-south-east"
  | "road-north-south-west"
  | "road-north-east-west"
  | "road-north-south-east"
  | "road-south-east-west";